Add support of all day events
The all day events behaves weirdly on o365. When created from the API,
it is really picky on the values we pass, otherwise it ends up
displaying an extra day on the calendar. We must have a timezone that
matches the one used by the user on o365 (I assume, or it does not like
UTC...).